@@ -1,10 +1,10 @@
-import { cookies } from 'next/headers';
+import { cookies } from "next/headers";
 
-import { AppSidebar } from '@/components/app-sidebar';
-import { SidebarInset, SidebarProvider } from '@/components/ui/sidebar';
+import { AppSidebar } from "@/components/app-sidebar";
+import { SidebarInset, SidebarProvider } from "@/components/ui/sidebar";
 
-import { auth } from '../(auth)/auth';
-import Script from 'next/script';
+import { auth } from "../(auth)/auth";
+import Script from "next/script";
 
 export const experimental_ppr = true;
 
@@ -14,7 +14,7 @@ export default async function Layout({
   children: React.ReactNode;
 }) {
   const [session, cookieStore] = await Promise.all([auth(), cookies()]);
-  const isCollapsed = cookieStore.get('sidebar:state')?.value !== 'true';
+  const isCollapsed = cookieStore.get("sidebar_state")?.value !== "true";
 
   return (
     <>
